+++ src/ImmichFolderWatch.Core/Interfaces/IActivationStateStore.cs
@@ -1,8 +1,0 @@
-namespace ImmichFolderWatch.Core.Interfaces;
-
-public interface IActivationStateStore
-{
-    bool IsInitialVerificationCompleted();
-
-    void MarkInitialVerificationCompleted();
-}